requirements.txt
contourpy==1.0.7
cycler==0.11.0
fonttools==4.38.0
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.3
numpy==1.24.1
packaging==23.0
Pillow==9.4.0
pyparsing==3.0.9
PyQt6==6.4.0
PyQt6-Qt6==6.4.2
PyQt6-sip==13.4.0
python-dateutil==2.8.2
scikit-learn==1.2.1
scipy==1.10.0
six==1.16.0
threadpoolctl==3.1.0
